Switch to QR code based identity transfer
This solves the problem of how to transfer to clients that no longer have any
identities, and makes the transfer mechanic more intuitive; the receive dialog
automatically closes on receipt of an identity, and as per the previous commit,
the received identity is automatically selected.

This change also makes newly created extra identities be automatically
selected, improves the handling of invalid QR scan results, and hides menu
items that are not relevent when a client has no identities.