Update to build tools v23, SDK v23 and v23 support libraries
As per https://code.google.com/p/android/issues/detail?id=183362, the previous
LastTransferFragment nesting method wasn't actually supported, and as the
upgrade to the v23 support library broke it, this change also switches to a new
method that does work.

The minSdkVersion is set to 22 because at least ICS and JB are currently broken
due to nested fragment issues and/or the new ServerTrust BKS keystore being too
new. Some effort to add back support for these platforms will have to be made
later.