Fix two ServerConnection bugs: disconnect when connecting caused (perm.) error
The permanent error state could only be recovered from by restarting the process
 - this part was a regression introduced by 9e706c10d31d because reconnect
requests in the general failure and TLS error states were ignored.

There was an older bug too, which was that disconnection requests during a
connect resulted in hasFailed being set, and, post-9e706c10d31d, it also caused
the state to transition to the new general failure state. This was because
the call to cancel on the webSocketCall in disconnect results in a call to the
onFailure callback. The fix is to add special handling for calls to onFailure in
the disconnecting state. Additionally, there was a race condition whereby
onFailure would be called before disconnect had set the state to disconnecting,
which resulted in the first part of the fix not having any effect (as the state
was still connecting). Hence the relocation of updating state in disconnect to
before the cancel/close calls.