Maintain connection request from onStart to onStop, fix two request state bugs
Re-requesting connections in onStart ensures the progress is immediately visible
i.e. that the "The connection was closed" message is never visible. It also
simplifies the activity code.

The two bugs could each individually cause the connection/join state to get
stuck at disconnecting/quitting game if a re-request was made while in one of
those states.

The change to identity receipt result handling is to protect against the unlikely
case of the BoardGameActivity being destroyed due to lack of memory - this way
the identity receipt activity will still close correctly.