Revert "Replace uses of deprecated onAttach(Activity) fragment method"
This reverts commit b4ccd1c14e6f88bdabb1a280d9bfe1b00625385e.

This made dialog fragments do nothing on Lollipop and lower, because the new
onAttach method wasn't called, so no listeners were registered. This mistake
was made because I think I _had_ observed the new method get called on
fragment compats. The difference is that none of the dialog fragments changed
here are using the support library dialog fragment, hence it not being called.