Revert targetSdkVersion to 23
Slightly desperate attempt to workaround crashes that occur when
targetSdkVersion >= 24 but which have no clear cause or fix, and which
there is no clear communication on from Google.

See also:
https://issuetracker.google.com/issues/37103380
https://www.devsbedevin.com/avoiding-transactiontoolargeexception-on-android-nougat-and-up/
https://github.com/guardian/toolargetool

Given that the four or five direct uses of onSaveInstance within
Liquidity are all with trivial payloads, my current working theory is
that the bug is in a 3rd party dependency. This is all conjecture, but
with a high number of crash reports in production and no better option,
this has to be worth a shot. If nothing else it will give an extra data
point to work with (i.e. confirmation or invalidation of this theory).