Generate fresh JWTs for each request
The observable of fresh JWTs has empirically proven to _not_ always be
fresh. Several times today I've observed failure of the new client to
connect, all of which were confirmed (via inspection of the server logs)
to be due to 401s thanks to the tokens not actually being fresh enough.

All such cases happened after the device had been idle for some time,
which provides a hint to the root cause.

Inspection of RxAndroid's HandlerScheduler quickly shows us how it makes
use of the Handler:

@Override
public Disposable scheduleDirect(Runnable run, long delay, TimeUnit unit) {
    if (run == null) throw new NullPointerException("run == null");
    if (unit == null) throw new NullPointerException("unit == null");

    run = RxJavaPlugins.onSchedule(run);
    ScheduledRunnable scheduled = new ScheduledRunnable(handler, run);
    handler.postDelayed(scheduled, Math.max(0L, unit.toMillis(delay)));
    return scheduled;
}

And as the docs for Handler.postDelayed tell us:

> Causes the Runnable r to be added to the message queue, to be run
after the specified amount of time elapses. The runnable will be run on
the thread to which this handler is attached. The time-base is
uptimeMillis(). Time spent in deep sleep will add an additional delay to
execution.

This could be mitigated by checking the freshness of the current value
and then explicitly creating a fresh one if the current is not, but that
would _add_ complexity (another option might be to reset the current
value after resuming from sleep).

Here we solve the problem by simply _removing_ complexity and generating
a fresh JWT for each call (and with tighter timing restrictions);
perhaps in a future revision command messages themselves will be signed
anyway.