Fix stuck-after-disconnect connections between onPause and onResume
source.exhausted() throws an exception when the call has been canceled,
and call.isCanceled was never returning true and thus not working as a
guard. The exception meant that onComplete was never called, but even if
it was, this wouldn't have worked because observables do not complete
once they have been disposed (the implementation filters any calls that
are made). Hence the correct thing being to update our state as soon as
we receive the call to quit(), which ensures we do then trigger a
reconnect when the activity resumes.