Move player and transfer diffing mechanisms from BoardGame to UI
In 2015 Android didn't provide all the mechanisms it now does for
differential RecyclerView updates. Now that it does, the diffing within
BoardGame itself is just unnecessary complexity. In addition, by
switching to ListAdapter, UI responsiveness should, if anything, be
improved, because the diffing is done by it on a separate thread.
BoardGame had been doing the work to provide equivalent behaviour via
collection diffing on the UI thread.