Fix LastTransferFragment regression
Caught via pre-launch testing report (the last two releases have been
internal-only).

The issue being that the synthetic textview_empty field is nulled when
the fragment is destroyed, hence the need to avoid running the update
callback in that case (before the callback was removed when the fragment
stopped but that was no longer done in order to avoid having a mutable
reference to the callback - the necessity of still removing it and the
possibility of doing so via a lifecycle observer having been
overlooked).

FATAL EXCEPTION: main
Process: com.dhpcs.liquidity, PID: 23731
java.lang.IllegalStateException: textview_empty must not be null
	at com.dhpcs.liquidity.fragment.LastTransferFragment.showTransfer(LastTransferFragment.java:71)
	at com.dhpcs.liquidity.fragment.LastTransferFragment.access$showTransfer(LastTransferFragment.java:23)
	at com.dhpcs.liquidity.fragment.LastTransferFragment$showTransfer$1.run(LastTransferFragment.java:104)
	at android.os.Handler.handleCallback(Handler.java:739)
	at android.os.Handler.dispatchMessage(Handler.java:95)
	at android.os.Looper.loop(Looper.java:135)
	at android.app.ActivityThread.main(ActivityThread.java:5538)
	at java.lang.reflect.Method.invoke(Method.java)
	at java.lang.reflect.Method.invoke(Method.java:372)
	at com.android.internal.os.ZygoteInit$MethodAndArgsCaller.run(ZygoteInit.java:960)
	at com.android.internal.os.ZygoteInit.main(ZygoteInit.java:755)