Fix fix fix LastTransferFragment regression
The callback was still getting run in a scenario where the synthetic
textview_empty field had been nulled (this is the problem with "magic"
like that - there's no documentation on the lifecycle of it, and no
implementation code to inspect - and hence no good leads to follow in
figuring out what the fix would be as long as synthetic view fields and
handlers are in).

The continued use of handler.post despite the otherwise wholesale move
to RxJava etc. was always a bit of a smell, and the continued existence
of the bug that results from messages not being removed from the handler
in the right order relative to textview_empty being nulled has acted as
the prompt to find a better way - which is what this commit does; we now
use Observable.interval and do the whole thing via Observables.